Update index.html
I filled in the information, such as my name for authors, brief description of the goal of the code, and known issues. I then created a function that will display my results by making variables for the percentages of the allergy and cold counts, variable for my results, and the function that will showcase my results. Lastly, I created a results page that contains the button to activate these functions, that will result in the display of my results.